Upgrade to subversion-1.7.3

+++ src/subversion/subversion/libsvn_subr/macos_keychain.c
@@ -2,17 +2,22 @@
  * macos_keychain.c: Mac OS keychain providers for SVN_AUTH_*
  *
  * ====================================================================
- * Copyright (c) 2003-2006, 2008 CollabNet.  All rights reserved.
+ *    Licensed to the Apache Software Foundation (ASF) under one
+ *    or more contributor license agreements.  See the NOTICE file
+ *    distributed with this work for additional information
+ *    regarding copyright ownership.  The ASF licenses this file
+ *    to you under the Apache License, Version 2.0 (the
+ *    "License"); you may not use this file except in compliance
+ *    with the License.  You may obtain a copy of the License at
  *
- * This software is licensed as described in the file COPYING, which
- * you should have received as part of this distribution.  The terms
- * are also available at http://subversion.tigris.org/license-1.html.
- * If newer versions of this license are posted there, you may use a
- * newer version instead, at your option.
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
- * This software consists of voluntary contributions made by many
- * individuals.  For exact contribution history, see the revision
- * history and logs, available at http://subversion.tigris.org/.
+ *    Unless required by applicable law or agreed to in writing,
+ *    software distributed under the License is distributed on an
+ *    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *    KIND, either express or implied.  See the License for the
+ *    specific language governing permissions and limitations
+ *    under the License.
  * ====================================================================
  */
 
@@ -75,25 +80,25 @@ keychain_password_set(apr_hash_t *creds,
   if (non_interactive)
     SecKeychainSetUserInteractionAllowed(FALSE);
 
-  status = SecKeychainFindGenericPassword(NULL, strlen(realmstring),
+  status = SecKeychainFindGenericPassword(NULL, (int) strlen(realmstring),
                                           realmstring, username == NULL
                                             ? 0
-                                            : strlen(username),
+                                            : (int) strlen(username),
                                           username, 0, NULL, &item);
   if (status)
     {
       if (status == errSecItemNotFound)
-        status = SecKeychainAddGenericPassword(NULL, strlen(realmstring),
+        status = SecKeychainAddGenericPassword(NULL, (int) strlen(realmstring),
                                                realmstring, username == NULL
                                                  ? 0
-                                                 : strlen(username),
-                                               username, strlen(password),
+                                                 : (int) strlen(username),
+                                               username, (int) strlen(password),
                                                password, NULL);
     }
   else
     {
       status = SecKeychainItemModifyAttributesAndData(item, NULL,
-                                                      strlen(password),
+                                                      (int) strlen(password),
                                                       password);
       CFRelease(item);
     }
@@ -122,10 +127,10 @@ keychain_password_get(const char **password,
   if (non_interactive)
     SecKeychainSetUserInteractionAllowed(FALSE);
 
-  status = SecKeychainFindGenericPassword(NULL, strlen(realmstring),
+  status = SecKeychainFindGenericPassword(NULL, (int) strlen(realmstring),
                                           realmstring, username == NULL
                                             ? 0
-                                            : strlen(username),
+                                            : (int) strlen(username),
                                           username, &length, &data, NULL);
 
   if (non_interactive)
